requirements.txt
# Core Django - Nâng cấp cho Python 3.10+
Django>=4.1.0,<4.3  # Django 4.1+ hỗ trợ Python 3.10
djangorestframework>=3.14.0
djangorestframework-simplejwt>=4.8.0
djoser>=2.1.0
django-cors-headers>=3.14.0
django-templated-mail>=1.1.1
django-filter>=23.2
django-storages>=1.13.2

# Database
dj-database-url>=1.2.0
psycopg2-binary>=2.9.10
pymysql>=1.1.0

# WebSocket & Cache
channels>=4.0.0
channels-redis>=4.1.0
redis>=4.6.0
django-redis>=5.3.0
daphne>=4.0.0

# Production server
gunicorn>=21.2.0
whitenoise>=6.5.0

# Image processing
Pillow>=12.0.0
cloudinary>=1.36.0
django-cloudinary-storage>=0.3.0

# Payment
stripe>=5.2.0

# AI Search (optional - có fallback nếu không có)
# torch>=2.0.0  # Uncomment nếu cần AI search
# transformers>=4.21.0
# sentence-transformers>=2.2.0

# Utilities
python-dotenv>=1.0.0
requests>=2.32.3

# Auth dependencies
PyJWT>=2.6.0
cryptography>=43.0.1
oauthlib>=3.2.2
requests-oauthlib>=1.3.1
social-auth-app-django>=4.0.0
social-auth-core>=4.3.0
python3-openid>=3.2.0

# Other dependencies
asgiref>=3.6.0
defusedxml>=0.7.1
sqlparse>=0.4.3
pytz>=2022.7.1
